+++ Sources/AutoSDK/AutoSDKVersion.m
@@ -1,4 +1,4 @@
 #import "include/AutoSDK.h"
 
-double AutoSDKVersionNumber = 1.15;
-const unsigned char AutoSDKVersionString[] = "1.15.0";
+double AutoSDKVersionNumber = 1.16;
+const unsigned char AutoSDKVersionString[] = "1.16.0";